Moving chart functions to charts.py from app.py
